backend/requirements.txt
# RealBoost AI — Backend Dependencies
# Install: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI
openai==1.52.0

# Payments
stripe==11.2.0

# Communications
twilio==9.3.6
sendgrid==6.11.0

# HTTP client
httpx==0.27.2

# Config
python-dotenv==1.0.1

# Utilities
pydantic[email]==2.9.2
